beauty_bot_function/requirements.txt
flask
requests
python-dotenv
beautifulsoup4
google-cloud-storage
google-cloud-logging
functions-framework==3.4.0